Remove interceptors for error handling due to lack of support in testing. Split browser interceptor into own file
